Refactor game-related tauri dunctions to frontend_api folder
