+++ src-tauri/src/frontend_api.rs
@@ -1,0 +1,1 @@
+pub mod games;
